requirements.txt
altgraph==0.17.3
CoolProp==6.4.3.post1
macholib==1.16.2
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
pyzmq==25.1.0
